+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.3 (2013/10/22)
+
+- Support `provision` command.
+
 # 0.0.2 (2013/10/08)
 
 - Add `vagrant sakura-list-id` command.